consolidate folders containing a single file
